requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
